# v2/frontend/src/api/generated/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Epic Controle V2",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/payments": {
      "post": {
        "tags": [
          "treasury"
        ],
        "summary": "Create Payment",
        "operationId": "create_payment_api_payments_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PaymentCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "treasury"
        ],
        "summary": "List Payments",
        "operationId": "list_payments_api_payments_get",
        "parameters": [
          {
            "name": "supplier_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Supplier Id"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "unallocated_only",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Unallocated Only"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PaymentResponse"
                  },
                  "title": "Response List Payments Api Payments Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/payments/with-document": {
      "post": {
        "tags": [
          "treasury"
        ],
        "summary": "Create Payment With Document",
        "operationId": "create_payment_with_document_api_payments_with_document_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_create_payment_with_document_api_payments_with_document_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/payments/{payment_id}": {
      "get": {
        "tags": [
          "treasury"
        ],
        "summary": "Get Payment",
        "operationId": "get_payment_api_payments__payment_id__get",
        "parameters": [
          {
            "name": "payment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Payment Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/payments/{payment_id}/eligible-payables": {
      "get": {
        "tags": [
          "treasury"
        ],
        "summary": "Eligible Payables",
        "operationId": "eligible_payables_api_payments__payment_id__eligible_payables_get",
        "parameters": [
          {
            "name": "payment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Payment Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EligiblePayable"
                  },
                  "title": "Response Eligible Payables Api Payments  Payment Id  Eligible Payables Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/payments/{payment_id}/allocations": {
      "post": {
        "tags": [
          "treasury"
        ],
        "summary": "Allocate",
        "operationId": "allocate_api_payments__payment_id__allocations_post",
        "parameters": [
          {
            "name": "payment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Payment Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AllocateBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/payments/{payment_id}/cancel": {
      "post": {
        "tags": [
          "treasury"
        ],
        "summary": "Cancel Payment",
        "operationId": "cancel_payment_api_payments__payment_id__cancel_post",
        "parameters": [
          {
            "name": "payment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Payment Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CancelBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AllocateBody": {
        "properties": {
          "expected_version": {
            "type": "integer",
            "title": "Expected Version"
          },
          "idempotency_key": {
            "type": "string",
            "title": "Idempotency Key"
          },
          "allocations": {
            "items": {
              "$ref": "#/components/schemas/AllocationLine"
            },
            "type": "array",
            "title": "Allocations"
          }
        },
        "type": "object",
        "required": [
          "expected_version",
          "idempotency_key",
          "allocations"
        ],
        "title": "AllocateBody"
      },
      "AllocationLine": {
        "properties": {
          "payable_id": {
            "type": "integer",
            "title": "Payable Id"
          },
          "amount": {
            "type": "string",
            "title": "Amount"
          },
          "expected_version": {
            "type": "integer",
            "title": "Expected Version"
          }
        },
        "type": "object",
        "required": [
          "payable_id",
          "amount",
          "expected_version"
        ],
        "title": "AllocationLine"
      },
      "AllocationResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "payable_id": {
            "type": "integer",
            "title": "Payable Id"
          },
          "amount": {
            "type": "string",
            "title": "Amount"
          },
          "batch_id": {
            "type": "integer",
            "title": "Batch Id"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "payable_id",
          "amount",
          "batch_id"
        ],
        "title": "AllocationResponse"
      },
...
      "Body_create_payment_with_document_api_payments_with_document_post": {
        "properties": {
          "supplier_id": {
            "type": "integer",
            "title": "Supplier Id"
          },
          "amount": {
            "type": "string",
            "title": "Amount"
          },
          "currency": {
            "type": "string",
            "title": "Currency",
            "default": "EUR"
          },
          "payment_date": {
            "type": "string",
            "format": "date",
            "title": "Payment Date"
          },
          "external_reference": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "External Reference"
          },
          "idempotency_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Idempotency Key"
          },
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "supplier_id",
          "amount",
          "payment_date",
          "file"
        ],
        "title": "Body_create_payment_with_document_api_payments_with_document_post"
      },
...
      "CancelBody": {
        "properties": {
          "expected_version": {
            "type": "integer",
            "title": "Expected Version"
          },
          "reason_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason Code"
          }
        },
        "type": "object",
        "required": [
          "expected_version"
        ],
        "title": "CancelBody"
      },
...
      "EligiblePayable": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "invoice_id": {
            "type": "integer",
            "title": "Invoice Id"
          },
          "order_id": {
            "type": "integer",
            "title": "Order Id"
          },
          "sequence": {
            "type": "integer",
            "title": "Sequence"
          },
          "due_date": {
            "type": "string",
            "format": "date",
            "title": "Due Date"
          },
          "amount": {
            "type": "string",
            "title": "Amount"
          },
          "balance": {
            "type": "string",
            "title": "Balance"
          },
          "allocated_amount": {
            "type": "string",
            "title": "Allocated Amount"
          },
          "currency": {
            "type": "string",
            "title": "Currency"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          }
        },
        "type": "object",
        "required": [
          "id",
          "invoice_id",
          "order_id",
          "sequence",
          "due_date",
          "amount",
          "balance",
          "allocated_amount",
          "currency",
          "status",
          "version"
        ],
        "title": "EligiblePayable"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PaymentCreate": {
        "properties": {
          "supplier_id": {
            "type": "integer",
            "title": "Supplier Id"
          },
          "amount": {
            "type": "string",
            "title": "Amount"
          },
          "currency": {
            "type": "string",
            "title": "Currency",
            "default": "EUR"
          },
          "payment_date": {
            "type": "string",
            "format": "date",
            "title": "Payment Date"
          },
          "external_reference": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "External Reference"
          },
          "idempotency_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Idempotency Key"
          },
          "register_without_document": {
            "type": "boolean",
            "title": "Register Without Document",
            "default": false
          },
          "reason_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason Code"
          }
        },
        "type": "object",
        "required": [
          "supplier_id",
          "amount",
          "payment_date"
        ],
        "title": "PaymentCreate"
      },
      "PaymentResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "supplier_id": {
            "type": "integer",
            "title": "Supplier Id"
          },
          "amount": {
            "type": "string",
            "title": "Amount"
          },
          "currency": {
            "type": "string",
            "title": "Currency"
          },
          "payment_date": {
            "type": "string",
            "format": "date",
            "title": "Payment Date"
          },
          "external_reference": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "External Reference"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "created_by_actor_id": {
            "type": "string",
            "title": "Created By Actor Id"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "idempotency_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Idempotency Key"
          },
          "register_without_document": {
            "type": "boolean",
            "title": "Register Without Document",
            "default": false
          },
          "amount_allocated": {
            "type": "string",
            "title": "Amount Allocated"
          },
          "amount_unallocated": {
            "type": "string",
            "title": "Amount Unallocated"
          },
          "allocations": {
            "items": {
              "$ref": "#/components/schemas/AllocationResponse"
            },
            "type": "array",
            "title": "Allocations"
          },
          "documents": {
            "items": {
              "$ref": "#/components/schemas/app__treasury__routes__DocumentBrief"
            },
            "type": "array",
            "title": "Documents"
          },
          "cancelled_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cancelled At"
          },
          "cancel_reason_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cancel Reason Code"
          }
        },
        "type": "object",
        "required": [
          "id",
          "supplier_id",
          "amount",
          "currency",
          "payment_date",
          "external_reference",
          "status",
          "created_by_actor_id",
          "version",
          "amount_allocated",
          "amount_unallocated"
        ],
        "title": "PaymentResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "app__treasury__routes__DocumentBrief": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "original_filename": {
            "type": "string",
            "title": "Original Filename"
          }
        },
        "type": "object",
        "required": [
          "id",
          "original_filename"
        ],
        "title": "DocumentBrief"
      }
    }
  }
}